Remove KindsImplicit module entirely
