+++ frontend/tests/cases/negative/implicitKind.gr.output
@@ -1,3 +1,3 @@
 Type checking failed: 
 Unification failed: frontend/tests/cases/negative/implicitKind.gr:10:1:
-Trying to unify `↑Nat` and `Type`
+Trying to unify `Type` and `↑Nat`
